we don't remove jabber avatars when it's not needed
git-svn-id: http://svn.miranda-ng.org/main/trunk@2722 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/JabberG/src/jabber_iqid.cpp
@@ -708,16 +708,12 @@ LBL_Ret:
 			}
 			if (item != NULL) {
 				hasPhoto = TRUE;
-				if (item->photoFileName)
+				if (item->photoFileName && _tcscmp(item->photoFileName, szAvatarFileName))
 					DeleteFile(item->photoFileName);
 				replaceStrT(item->photoFileName, szAvatarFileName);
 				Log("Contact's picture saved to " TCHAR_STR_PARAM, szAvatarFileName);
-
-				if (JGetWord(hContact, "Status", ID_STATUS_OFFLINE) == ID_STATUS_OFFLINE) {
-					char szHashValue[ MAX_PATH ];
-					if (JGetStaticString("AvatarHash", hContact, szHashValue, sizeof(szHashValue)))
-						OnIqResultGotAvatar(hContact, o, xmlGetText(m));
-			}	}
+				OnIqResultGotAvatar(hContact, o, xmlGetText(m));
+			}
 
 			db_free(&dbv);
 	}	}
